Fix kubernetes upgrades
This change does the following:
 - Upgrades kops to version 1.14.0. New kubernetes clusters are
   now created with kubernetes 1.14.x.
 - Fix kops provisioner upgrade flow. A few small bugs were fixed
   in order to provide for the ability to upgrade clusters. The
   functionality remains the same for now where clusters are
   upgraded to the latest "stable" kops release when the command
   is issued.
 - Add a few more kops and terraform commands to support the
   kops upgrade flow.
