+++ Makefile
@@ -8,7 +8,7 @@ BUILD_TIME := $(shell date -u +%Y%m%d.%H%M%S)
 BUILD_HASH := $(shell git rev-parse HEAD)
 
 TERRAFORM_VERSION=0.11.14
-KOPS_VERSION=1.13.2
+KOPS_VERSION=1.14.0
 HELM_VERSION=v2.14.1
 KUBECTL_VERSION=v1.14.0
